watcher/requirements.txt
watchdog==6.0.0
psycopg2-binary==2.9.10
